+++ Sources/VibeToken/Domain/LaunchAtLogin.swift
@@ -1,0 +1,22 @@
+import ServiceManagement
+
+enum LaunchAtLogin {
+    static var isRegistered: Bool {
+        switch SMAppService.mainApp.status {
+        case .enabled, .requiresApproval:
+            return true
+        case .notRegistered, .notFound:
+            return false
+        @unknown default:
+            return false
+        }
+    }
+
+    static func setRegistered(_ registered: Bool) throws {
+        if registered {
+            try SMAppService.mainApp.register()
+        } else {
+            try SMAppService.mainApp.unregister()
+        }
+    }
+}